New Card WIP and New option for slider card
Started creating a footer card, it will be somewhat customizable, but mainly just because I don't wanna do it with a HA Markdown card.

I have also added a new config key 'toggle_scroll'. If this is set to true, then it will disable scrolling for the whole window. This is also just something for me, I don't know if anybody will find it useful. I have vertical cover sliders, and using them on the phone sucks. This has eliminated that issue, but the scrollbar on pc will appear/diappear when using the slider. Thats why it's optional.