Created two new cards that are ready to be tested. my-button-light & my-button-cover.
These cards are customizable to some extend as they are now. With this I mean, that the styling can pretty much be customized how you like, except for the slider track and thumb, since they cannot be styled with inline styles. But I have thought of a way it might be able to be implemented so those are also 100% customizable.

The cards has pretty much been set up to fit my needs, but they should be usable for others. Probably more so than the standalone slider card. (I promise I will look into some of the issues. the vertical slider will probably not become much better than it is, becuase there are some css attributes I use to hide the default look of a slider. If I have to make it vertical, those same css attributes needs to be set to vertical-slider, which than also shows the ugly default slider track. So it has to be with rotation. Please if there are any front-end devs out there to tell me otherwise, I would love that.)

I will write a readme on the two new cards with a couple of demonstration photos and yaml configs.

I will probably look into updating the my-slider soon as well. I will be creating a new config variable named something like 'version'. And the default will be version 1, which is the current version, and then people can choose to set version as 2 instead, which will use version 2 of the card. This way, noone gets mad like the last time I made major unannounced changes to the card. Because I will be making it so styles can be set 100% customizable like with the new cards (also track and thumb). and some of the config keys will probably be renamed to make more sense.
Hopefully I will also be able to look into text in the slider container. It has to be possible. I just doubt it will be with jinja templating, so it will probably start out with jut being a couple of config variables to tell the card what to show. For example only text or text and light percentage.