I have started creating a brand new slider v2. This will not be using the input range html element, but simply be made out of divs and be handled by javascript. This way I have full control of everything and you guys can easily get full control of all the styling. And this way we can easily make vertical sliders.
I don't know when it'll be ready. But it will not interfere with the old slider card and it's config. I will make it a completely new card 'my-slider-v2'. The old card will become depcrecated when this is released and eventually it will be removed completely from the compiled script. But that is long in the future.