my-slider-v2 is looking mighty fine by now. I should have made my own slider a long time ago. So far it seems like more work, but much more control of the components and such. So far I haven't stumpled upon any major or minor bugs. The color_temp one is a bit dodgy, but I have never used that before, so maybe it has always been weird with updating the progress sporadically.
I have implemented most of the logic for the different entity types. Still missing a couple though. I just wanted to commit at this stage, because I have a pretty solid base to continue from here. I have just added the config key 'showMin'. This will now make it so we can have custom min/max and we can decide to show the min/max on the slider or the slider can be the min/max. Thinking about it, I actually don't know if it will work mith max. Or should I say, max will always be the far right.. Right?? :)

But yeah, making good progress on it and it will be released with an updated readme soon.