allowSliding and slideDistance added to my-slider-v2
allowSliding will allow the user to slide in the direction of the slider for an amount of pixels and start sliding. This is in conjuction with allowTapping false so you can start sliding anywhere on the slider card and when you surpass the distance it will activate.
slideDistance is the other new config key. This is default 10. It determines the amount of pixels the input has the travel for allowSliding to take effect.